revert unnecessary send_request-factory
