+++ lib/mixpanel-node.js
@@ -7,21 +7,30 @@
     Released under the MIT license.
 */
 
-const {async_all, ensure_timestamp, create_config, create_send_request_func} = require('./utils');
+const querystring = require('querystring');
+const Buffer = require('buffer').Buffer;
+const http = require('http');
+const https = require('https');
+const HttpsProxyAgent = require('https-proxy-agent');
+
+const {async_all, ensure_timestamp, create_config} = require('./utils');
 const {create_profile_helpers} = require('./profile_helpers');
 const {MixpanelGroups} = require('./groups');
 const {MixpanelPeople} = require('./people');
 
 var create_client = function(token, config) {
-    // mixpanel constants
-    const MAX_BATCH_SIZE = 50;
-    const TRACK_AGE_LIMIT = 60 * 60 * 24 * 5;
-
     if (!token) {
         throw new Error("The Mixpanel Client needs a Mixpanel token: `init(token)`");
     }
 
-    let metrics = {token};
+    // mixpanel constants
+    const MAX_BATCH_SIZE = 50;
+    const TRACK_AGE_LIMIT = 60 * 60 * 24 * 5;
+    const REQUEST_LIBS = {http, https};
+    const proxyPath = process.env.HTTPS_PROXY || process.env.HTTP_PROXY;
+    const proxyAgent = proxyPath ? new HttpsProxyAgent(proxyPath) : null;
+
+    const metrics = {token};
 
     /**
         set_config(config)
@@ -33,14 +42,116 @@ var create_client = function(token, config) {
     */
     metrics.set_config = (new_config) => {
         const config = create_config(new_config);
-        const send_request = create_send_request_func(config);
-        const profile_helpers = create_profile_helpers({token: metrics.token, config, send_request});
-        Object.assign(metrics, {config, send_request, _helpers: profile_helpers});
+        const profile_helpers = create_profile_helpers({token: metrics.token, config, send_request: metrics.send_request});
+        Object.assign(metrics, {config, _helpers: profile_helpers});
     };
     metrics.set_config(config);
     metrics.groups = new MixpanelGroups(metrics);
     metrics.people = new MixpanelPeople(metrics);
 
+
+    /**
+     * sends an async GET or POST request to mixpanel
+     * for batch processes data must be send in the body of a POST
+     * @param {object} options
+     * @param {string} options.endpoint
+     * @param {object} options.data         the data to send in the request
+     * @param {string} [options.method]     e.g. `get` or `post`, defaults to `get`
+     * @param {function} callback           called on request completion or error
+     */
+    metrics.send_request = function(options, callback) {
+        callback = callback || function() {};
+
+        let content = Buffer.from(JSON.stringify(options.data)).toString('base64');
+        const endpoint = options.endpoint;
+        const method = (options.method || 'GET').toUpperCase();
+        let query_params = {
+            'ip': 0,
+            'verbose': metrics.config.verbose ? 1 : 0
+        };
+        const key = metrics.config.key;
+        const request_lib = REQUEST_LIBS[metrics.config.protocol];
+        let request_options = {
+            host: metrics.config.host,
+            port: metrics.config.port,
+            headers: {},
+            method: method
+        };
+        let request;
+
+        if (!request_lib) {
+            throw new Error(
+                "Mixpanel Initialization Error: Unsupported protocol " + metrics.config.protocol + ". " +
+                "Supported protocols are: " + Object.keys(REQUEST_LIBS)
+            );
+        }
+
+
+        if (method === 'POST') {
+            content = 'data=' + content;
+            request_options.headers['Content-Type'] = 'application/x-www-form-urlencoded';
+            request_options.headers['Content-Length'] = Buffer.byteLength(content);
+        } else if (method === 'GET') {
+            query_params.data = content;
+        }
+
+
+        // add `key` query params
+        if (key) {
+            query_params.api_key = key;
+        } else if (endpoint === '/import') {
+            throw new Error("The Mixpanel Client needs a Mixpanel api key when importing old events: `init(token, { key: ... })`");
+        }
+
+        if (proxyAgent) {
+            request_options.agent = proxyAgent;
+        }
+
+        if (metrics.config.test) {
+            query_params.test = 1;
+        }
+
+        request_options.path = metrics.config.path + endpoint + "?" + querystring.stringify(query_params);
+
+        request = request_lib.request(request_options, function(res) {
+            var data = "";
+            res.on('data', function(chunk) {
+                data += chunk;
+            });
+
+            res.on('end', function() {
+                var e;
+                if (metrics.config.verbose) {
+                    try {
+                        var result = JSON.parse(data);
+                        if(result.status != 1) {
+                            e = new Error("Mixpanel Server Error: " + result.error);
+                        }
+                    }
+                    catch(ex) {
+                        e = new Error("Could not parse response from Mixpanel");
+                    }
+                }
+                else {
+                    e = (data !== '1') ? new Error("Mixpanel Server Error: " + data) : undefined;
+                }
+
+                callback(e);
+            });
+        });
+
+        request.on('error', function(e) {
+            if (metrics.config.debug) {
+                console.log("Got Error: " + e.message);
+            }
+            callback(e);
+        });
+
+        if (method === 'POST') {
+            request.write(content);
+        }
+        request.end();
+    };
 
     /**
      * Send an event to Mixpanel, using the specified endpoint (e.g., track/import)
